#7- 문제 세팅
