+++ Algorithm/Algorithm/BOJ/main.swift
@@ -1,19 +1,10 @@
 //
-//  2562.swift
-//  BOJ_1264.swift
+//  2601.swift
+//  약수 구하기
 //  Algorithm
 //
-//  Created by 김인영 on 2023/09/03.
+//  Created by 김인영 on 2023/09/06.
 //
 
 import Foundation
 
-var numbers: [Int] = []
-
-for _ in 1...9 {
-    numbers.append(Int(readLine()!)!)
-}
-
-let maximum = numbers.max()!
-let indexOfMaximum = numbers.firstIndex(of: maximum)! + 1
-print("\(maximum)\n" + "\(indexOfMaximum)")
